+++ packages/create-app/template/+next+fuma-docs-mdx/lib/source.ts
@@ -21,7 +21,7 @@ export function getPageImage(page: InferPageType<typeof source>) {
 export async function getLLMText(page: InferPageType<typeof source>) {
   const processed = await page.data.getText('processed');
 
-  return `# ${page.data.title} (${page.url})
+  return `# ${page.data.title}
 
 ${processed}`;
 }