+++ packages/create-app/template/+orama-cloud/+next+fuma-docs-mdx/app/layout.tsx
@@ -1,0 +1,18 @@
+import { RootProvider } from 'fumadocs-ui/provider/next';
+import './global.css';
+import { Inter } from 'next/font/google';
+import SearchDialog from '@/components/search';
+
+const inter = Inter({
+  subsets: ['latin'],
+});
+
+export default function Layout({ children }: LayoutProps<'/'>) {
+  return (
+    <html lang="en" className={inter.className} suppressHydrationWarning>
+      <body className="flex flex-col min-h-screen">
+        <RootProvider search={{ SearchDialog }}>{children}</RootProvider>
+      </body>
+    </html>
+  );
+}