+++ packages/create-app/template/+orama-cloud/waku/src/components/provider.tsx
@@ -1,0 +1,8 @@
+'use client';
+import type { ReactNode } from 'react';
+import { RootProvider } from 'fumadocs-ui/provider/waku';
+import SearchDialog from '@/components/search';
+
+export function Provider({ children }: { children: ReactNode }) {
+  return <RootProvider search={{ SearchDialog }}>{children}</RootProvider>;
+}