+++ packages/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # next-docs-zeta
 
+## 16.0.2
+
+### Patch Changes
+
+- d511232: Fix i18n middleware search params handling
+
+## 16.0.1
+
+### Patch Changes
+
+- 45f0c1f: hotfix `<DynamicCodeBlock />` Vite + React 19.2 compat issues
+
 ## 16.0.0
 
 ### Major Changes